web更好的连接检测
